projectx/requirements.txt
requests>=2.28
PyYAML>=6.0
pandas>=1.3
numpy>=1.20
python-dotenv>=1.0
signalrcore>=1.0.2
websocket-client>=1.6
